movie_info: change the way get_movie_info works to match more accurately
Before this patch, get_movie_info was just searching for an exact match
in the title, without even looking if the result was a movie or a series.
With this patch, if there is no result at first or if the result is a
tv show, get_movie_info will use the omdbapi to search for results and,
if there are some, order the results by proximity with the request before
using the best match (if proximity >= 90%). This allows to have some
errors in the file name.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>