TraktClient: add PIN auth
Trakt.tv changed its v2 API to introduce a new PIN authentication
mode. With this new mode, the token authentication is now
deprecated, thus still usable for the moment. This commit brings
changes to TraktClient.py to allow both authentication mode
depending on how we open the session at first. It also brings
changes to TraktForVLC.py to use the new initialization of the
TraktClient module and read and update the new config options
(PIN as well as access and refresh tokens).

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>