Fix: remove OSX from Travis CI configuration
Travis CI has enable multi platform support for all accounts
without discarding the ones that are not available. Python
is not yet available for OSX as mentionned in [1]. It will
thus be added again when it will be supported, but removed
for now from the platforms to use to run tests.

[1]: https://github.com/travis-ci/travis-ci/issues/4729#issuecomment-136519696

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>