ci: Update the CI processes and build
CIs will now use the requirements.txt file to install dependencies,
and AppVeyor has been configured to generate and publish an .exe
file off the application containing all dependencies when building
a tag. This means that user will not need to install everything
anymore, but instead just download a zip.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>