ci: re-enable tests and allow them to be ran on OS X
Tests were disabled previously because of a regular fail using
imdbpie. This commit introduces a retry scheme for movie_info
that allows to retry when receiving a 503 error from imdb. This
kind of retry should never happen in normal use cases but is
coherent with the number of requests made by a CI.

Tests should be updated in the future to avoid those heavy loads
on external services.

Support for Python 2.6 has also been dropped.

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>