helper: commands: resolve: fix hash match when multiple entries are returned
When multiple entries are returned by the OpenSubtitles API, we
first check for episodes and then movies. However, nothing
guarantees that both of those will have data in them. We thus
need to first compute the list of matching returned medias of
the right type, and then, if any matches, determine the closest.

Fixes #96

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>