helper: commands: resolve: manage mathematical issue for movie resolution
In some cases, we can get mean + stddev > max, in such cases, the
movie resolution was failing as it was looking for any movie having
a match of at least mean + stddev. Using a "min" operator with the
"max" value available for the movies, we insure that we will always
at least have a match.

Fixes #98

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>