helper: commands: resolve: allows to resolve episodes without IMDb match
In some cases, episodes seem to be weirdly registered in IMDb (two
by two for instance) which means the matching between the resolution
done using The TVDB and the corresponding IMDb episode does not work.
This commit fixes that by allowing to resolve without an IMDb match
by querying directly Trakt if the match is not found. This also
introduces the use of other ids to perform the scrobble.

This commit also introduce a compatibility code in trakt.lua that
will need to be removed in a few versions.

Fixes #111

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>