intf: movieHash: manage file opening errors properly
When failing to open a file to resolve the hash, the interface was
exiting on error because of the assertion. It can happen for a
number of reasons that the file cannot be opened and this should
not prevent TraktForVLC from working properly. This commit aims
at managing the error properly by showing an error message and
aborting the hash resolution.

Relates to #115

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>