intf: Windows: manage Universal Naming Convention paths
When a network drive is not mapped to a drive letter on Windows,
the path will appear as an UNC path starting with 'file://'.
However, Lua will fail to open that file to resolve the hash.
This thus fixes that problem by changing the protocol from 'file'
to 'unc' in order to avoid trying to resolve the hash when
reading a media from an UNC path.

Relates to #115

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>