helper: commands: resolve: fix movie resolution
Commit f43646067521d794661b69b8a2b3aa372466d5a2 introduced a
bug by converting the imdb id returned in the URL form by
imdbpie to the imdb id alone, and doing so for the
'parentTitle' of the media, which represents a series for
a TV episode, but does not exist for a movie. This commit
fixes that behavior by checking that the 'parentTitle'
exists before applying the conversion at its level.

Fixes #119

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>